requirements.txt
torch>=2.3.0
torchvision>=0.18.0
matplotlib>=3.8.0
numpy>=1.26.0
pillow>=10.0.0
PyYAML>=6.0.1
scikit-learn>=1.5.0
